refactor: 커스텀 훅에서 쓰는 타입들을 src/types로 분리함
