+++ front/src/types/States.ts
@@ -14,3 +14,21 @@ export interface UserState {
   gender: number | null;
   age: number | null;
 }
+
+export interface RequestState<T> {
+  type: 'NULL' | 'REQUEST' | 'SUCCESS' | 'FAILURE';
+  requestProps?: any[];
+  response?: T;
+  statusCode?: number;
+}
+
+export interface UploadState {
+  willUploadFiles: File[];
+  presignedURLs: string[];
+}
+
+export interface WorldcupState {
+  title: string;
+  desc: string;
+  keywords: string[];
+}
